work 15/6/2025
Today I completed the code. I modified the battle file to fit the main and ui. In the character file, I added abilities to all the game classes. I also modified the file to fit the main and ui. For the main, I completed the game by adding more bosses and butchering it to look better. For the ui, I added the 'help' button which tells you what you must do in the game and I also changed the fonts/backgrounds for certain screens and text. Finally, I removed the requiremenst.txt because I already had everything I needed for the project.
The code is complete